+++ tt_metal/api/tt-metalium/command_queue_interface.hpp
@@ -9,7 +9,6 @@
 #include <tt-metalium/tt_align.hpp>
 #include <tt-metalium/fabric_host_interface.h>
 
-#include <tt-metalium/launch_message_ring_buffer_state.hpp>
 #include <tt-metalium/buffer.hpp>
 #include <umd/device/tt_core_coordinates.h>
