+++ tt_metal/api/tt-metalium/mesh_command_queue.hpp
@@ -22,7 +22,6 @@
 #include <tt-metalium/command_queue.hpp>
 #include <tt-metalium/command_queue_interface.hpp>
 #include <tt-metalium/core_coord.hpp>
-#include <tt-metalium/launch_message_ring_buffer_state.hpp>
 #include <tt-metalium/mesh_buffer.hpp>
 #include <tt-metalium/mesh_coord.hpp>
 #include <tt-metalium/mesh_device.hpp>
