+++ tt_metal/distributed/fd_mesh_command_queue.hpp
@@ -10,6 +10,7 @@
 #include <tt-metalium/multi_producer_single_consumer_queue.hpp>
 
 #include "dispatch/dispatch_settings.hpp"
+#include "dispatch/launch_message_ring_buffer_state.hpp"
 
 namespace tt::tt_metal::distributed {
