+++ tt_metal/distributed/mesh_device.cpp
@@ -26,7 +26,6 @@
 #include "allocator.hpp"
 #include "assert.hpp"
 #include "dispatch/dispatch_settings.hpp"
-#include "launch_message_ring_buffer_state.hpp"
 #include "mesh_trace.hpp"
 #include "shape_base.hpp"
 #include <tt_stl/span.hpp>
@@ -43,6 +42,7 @@
 
 #include "tt_metal/impl/allocator/l1_banking_allocator.hpp"
 #include "tt_metal/impl/sub_device/sub_device_manager.hpp"
+#include "dispatch/launch_message_ring_buffer_state.hpp"
 #include "sub_device/sub_device_manager_tracker.hpp"
 #include <umd/device/types/xy_pair.h>
