+++ tt_metal/impl/dispatch/launch_message_ring_buffer_state.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <launch_message_ring_buffer_state.hpp>
+#include "launch_message_ring_buffer_state.hpp"
 
 #include <cstdint>
